+++ src/Catalog/OfferClient.php
@@ -51,11 +51,11 @@ class OfferClient extends AbstractCatalogClient
      * @param string $sku The sku of the offer
      * @param array $parameters The parameters to add to the URL
      *
-     * @return string|null The response
+     * @return string The response
      *
      * @throws PublicApiClientException
      */
-    public function getOfferDetailsRaw(string $classification, string $vendorCode, string $sku, array $parameters = []): ?string
+    public function getOfferDetailsRaw(string $classification, string $vendorCode, string $sku, array $parameters = []): string
     {
         $classification = urlencode($classification);
         $vendorCode = urlencode($vendorCode);
@@ -103,10 +103,18 @@ class OfferClient extends AbstractCatalogClient
      * @param array $parameters Parameters to append to the query string
      *
      * @return string The response
+     * @throws NotFoundException
+     * @throws PublicApiClientException
      */
     public function findRaw(array $postData, array $parameters = []): string
     {
         $this->path = self::FIND_PATH;
+
+        if (isset($postData[self::DATA_FILTERS])) {
+            $postData[self::DATA_FILTERS] = array_map(static function ($row) {
+                return is_array($row) ? array_values($row) : $row;
+            }, $postData[self::DATA_FILTERS]);
+        }
 
         return $this->post($postData, $parameters);
     }