+++ src/PublicApiClient.php
@@ -4,6 +4,7 @@ namespace ArrowSphere\PublicApiClient;
 
 use ArrowSphere\PublicApiClient\Catalog\AddonClient;
 use ArrowSphere\PublicApiClient\Catalog\ClassificationClient;
+use ArrowSphere\PublicApiClient\Catalog\FamilyClient;
 use ArrowSphere\PublicApiClient\Catalog\OfferClient;
 use ArrowSphere\PublicApiClient\Catalog\ProgramClient;
 use ArrowSphere\PublicApiClient\Catalog\ServiceClient;
@@ -21,7 +22,7 @@ class PublicApiClient extends AbstractClient
      */
     public function getWhoamiClient(): WhoamiClient
     {
-        return (new WhoamiClient($this->curler))
+        return (new WhoamiClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
@@ -31,7 +32,7 @@ class PublicApiClient extends AbstractClient
      */
     public function getCheckDomainClient(): CheckDomainClient
     {
-        return (new CheckDomainClient($this->curler))
+        return (new CheckDomainClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
@@ -41,7 +42,7 @@ class PublicApiClient extends AbstractClient
      */
     public function getOfferClient(): OfferClient
     {
-        return (new OfferClient($this->curler))
+        return (new OfferClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
@@ -51,7 +52,7 @@ class PublicApiClient extends AbstractClient
      */
     public function getProgramClient(): ProgramClient
     {
-        return (new ProgramClient($this->curler))
+        return (new ProgramClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
@@ -61,7 +62,7 @@ class PublicApiClient extends AbstractClient
      */
     public function getAddonClient(): AddonClient
     {
-        return (new AddonClient($this->curler))
+        return (new AddonClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
@@ -71,17 +72,28 @@ class PublicApiClient extends AbstractClient
      */
     public function getClassificationClient(): ClassificationClient
     {
-        return (new ClassificationClient($this->curler))
+        return (new ClassificationClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
 
     /**
      * @return ServiceClient
+     * @deprecated use getFamilyClient() instead
      */
     public function getServiceClient(): ServiceClient
     {
-        return (new ServiceClient($this->curler))
+        return (new ServiceClient($this->client))
+            ->setUrl($this->url)
+            ->setApiKey($this->apiKey);
+    }
+
+    /**
+     * @return FamilyClient
+     */
+    public function getFamilyClient(): FamilyClient
+    {
+        return (new FamilyClient($this->curler))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }
@@ -91,7 +103,7 @@ class PublicApiClient extends AbstractClient
      */
     public function getLicensesClient(): LicensesClient
     {
-        return (new LicensesClient($this->curler))
+        return (new LicensesClient($this->client))
             ->setUrl($this->url)
             ->setApiKey($this->apiKey);
     }