+++ tests/Catalog/AddonClientTest.php
@@ -8,6 +8,7 @@ use ArrowSphere\PublicApiClient\Catalog\Entities\PriceBand;
 use ArrowSphere\PublicApiClient\Exception\NotFoundException;
 use ArrowSphere\PublicApiClient\Exception\PublicApiClientException;
 use ArrowSphere\PublicApiClient\Tests\AbstractClientTest;
+use GuzzleHttp\Psr7\Response;
 
 /**
  * Class AddonClientTest
@@ -24,11 +25,11 @@ class AddonClientTest extends AbstractClientTest
      */
     public function testGetAddonsRaw(): void
     {
-        $this->curler->response = 'OK USA';
-
-        $this->curler->expects(self::once())
-            ->method('get')
-            ->with('https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons');
+        $this->httpClient
+            ->expects(self::once())
+            ->method('request')
+            ->with('get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons')
+            ->willReturn(new Response(200, [], 'OK USA'));
 
         $this->client->getAddonsRaw('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483');
     }
@@ -40,7 +41,11 @@ class AddonClientTest extends AbstractClientTest
      */
     public function testGetAddonsWithInvalidResponse(): void
     {
-        $this->curler->response = '{';
+        $this->httpClient
+            ->expects(self::once())
+            ->method('request')
+            ->with('get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?per_page=100')
+            ->willReturn(new Response(200, [], '{'));
 
         $this->expectException(PublicApiClientException::class);
         $addons = $this->client->getAddons('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483');
@@ -54,20 +59,22 @@ class AddonClientTest extends AbstractClientTest
      */
     public function testGetAddonsWithPagination(): void
     {
-        $this->curler->response = json_encode([
+        $response = json_encode([
             'data'       => [],
             'pagination' => [
                 'total_page' => 3,
             ],
         ]);
 
-        $this->curler->expects(self::exactly(3))
-            ->method('get')
+        $this->httpClient
+            ->expects(self::exactly(3))
+            ->method('request')
             ->withConsecutive(
-                ['https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?per_page=100'],
-                ['https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?page=2&per_page=100'],
-                ['https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?page=3&per_page=100']
-            );
+                ['get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?per_page=100'],
+                ['get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?page=2&per_page=100'],
+                ['get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?page=3&per_page=100']
+            )
+            ->willReturn(new Response(200, [], $response));
 
         $test = $this->client->getAddons('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483');
         iterator_to_array($test);
@@ -80,7 +87,7 @@ class AddonClientTest extends AbstractClientTest
      */
     public function testGetAddons(): void
     {
-        $this->curler->response = <<<JSON
+        $response = <<<JSON
 {
   "status": 200,
   "data": [
@@ -277,9 +284,11 @@ class AddonClientTest extends AbstractClientTest
 }
 JSON;
 
-        $this->curler->expects(self::once())
-            ->method('get')
-            ->with('https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?per_page=100');
+        $this->httpClient
+            ->expects(self::once())
+            ->method('request')
+            ->with('get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons?per_page=100')
+            ->willReturn(new Response(200, [], $response));
 
         $test = $this->client->getAddons('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483');
         $list = iterator_to_array($test);
@@ -424,11 +433,11 @@ JSON;
      */
     public function testGetAddonRaw(): void
     {
-        $this->curler->response = 'OK USA';
-
-        $this->curler->expects(self::once())
-            ->method('get')
-            ->with('https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons/0F598EFE-F330-4D79-B79F-C9480BB7CE3E');
+        $this->httpClient
+            ->expects(self::once())
+            ->method('request')
+            ->with('get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons/0F598EFE-F330-4D79-B79F-C9480BB7CE3E')
+            ->willReturn(new Response(200, [], 'OK USA'));
 
         $this->client->getAddonRaw('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483', '0F598EFE-F330-4D79-B79F-C9480BB7CE3E');
     }
@@ -440,7 +449,11 @@ JSON;
      */
     public function testGetAddonWithInvalidResponse(): void
     {
-        $this->curler->response = '{';
+        $this->httpClient
+            ->expects(self::once())
+            ->method('request')
+            ->with('get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons/0F598EFE-F330-4D79-B79F-C9480BB7CE3E')
+            ->willReturn(new Response(200, [], '{'));
 
         $this->expectException(PublicApiClientException::class);
         $this->client->getAddon('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483', '0F598EFE-F330-4D79-B79F-C9480BB7CE3E');
@@ -453,7 +466,7 @@ JSON;
      */
     public function testGetAddon(): void
     {
-        $this->curler->response = <<<JSON
+        $response = <<<JSON
 {
   "status": 200,
   "data": {
@@ -548,10 +561,11 @@ JSON;
   }
 }
 JSON;
-
-        $this->curler->expects(self::once())
-            ->method('get')
-            ->with('https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons/0F598EFE-F330-4D79-B79F-C9480BB7CE3E');
+        $this->httpClient
+            ->expects(self::once())
+            ->method('request')
+            ->with('get', 'https://www.test.com/catalog/categories/SAAS/programs/microsoft/products/MS-1A-M365-ENT/offers/05933B83-04D1-4AC6-BFCD-DBBBFC834483/addons/0F598EFE-F330-4D79-B79F-C9480BB7CE3E')
+            ->willReturn(new Response(200, [], $response));
 
         $addon = $this->client->getAddon('SAAS', 'microsoft', 'MS-1A-M365-ENT', '05933B83-04D1-4AC6-BFCD-DBBBFC834483', '0F598EFE-F330-4D79-B79F-C9480BB7CE3E');
         self::assertEquals('description', $addon->getDescription());